make Quit main menu action work
The `connect` function may as well take 3 arguments, with new
syntax. Maybe with old too, though, but that's not really matters.

So, the first two are as usual, sender and address of signal
function, and the third is just an address of _static_ function, exactly
like `QAppliction::quit` is.

Also this commit rearranges order of creation of menu entries, to
represent their tree-like structure better (IMHO, of course, so this is
a subject to change at any time)